@@ -8,8 +8,9 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "axios": "^0.25.0",
     "core-js": "^3.6.5",
-    "vue": "^2.6.11"
+    "vue": "^3.2.29"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
