@@ -1,8 +1,7 @@
-import Vue from 'vue'
-import App from './App.vue'
+import {createApp} from 'vue'
+import App from './App'
+import components from '@/components/UI'
 
-Vue.config.productionTip = false
-
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+const app = createApp(App)
+components.forEach(component => app.component(component.name, component))
+app.mount('#app')
